add database dependency
let's use lowdb as the database for this project - it's not
production ready, but it expects the same behaviors from consumers (it's
async, needs to be properly opened / closed / flushed, etc).

this prevents needing to stand up a db server for local testing (and, as
an added benefit, it has a pleasant, functional API style).